+++ src/main/java/com/alibaba/ttl/threadpool/agent/TtlTransformer.java
@@ -1,5 +1,6 @@
 package com.alibaba.ttl.threadpool.agent;
 
+import com.alibaba.ttl.threadpool.agent.internal.transformlet.JavassistTransformlet;
 import javassist.ClassPool;
 import javassist.CtClass;
 import javassist.LoaderClassPath;
