--- src/main/java/com/alibaba/ttl/threadpool/agent/transformlet/package-info.java
@@ -1,7 +1,0 @@
-/**
- * TTL {@code Transformlet} implementations by {@code Javassist}.
- *
- * @author Jerry Lee (oldratlee at gmail dot com)
- * @see com.alibaba.ttl.threadpool.agent.JavassistTransformlet
- */
-package com.alibaba.ttl.threadpool.agent.transformlet;
